+++ test/Atata.WebDriverSetup.IntegrationTests/DriverSetupTests.cs
@@ -95,6 +95,31 @@ namespace Atata.WebDriverSetup.IntegrationTests
             AssertAutoSetUpDriverResults(results, browserNames?.Where(IsValidBrowserName) ?? new string[0]);
         }
 
+        [Test]
+        public async Task SetUpAsync_DifferentBrowsers_InParallel()
+        {
+            var results = await Task.WhenAll(
+                DriverSetup.ConfigureChrome().SetUpAsync(),
+                DriverSetup.ConfigureFirefox().SetUpAsync(),
+                DriverSetup.ConfigureEdge().SetUpAsync());
+
+            AssertAutoSetUpDriverResults(results, new[] { BrowserNames.Chrome, BrowserNames.Firefox, BrowserNames.Edge });
+        }
+
+        [Test]
+        public async Task SetUpAsync_SameBrowsers_WithMutex_InParallel()
+        {
+            var results = await Task.WhenAll(
+                DriverSetup.ConfigureChrome().WithMutex(true).SetUpAsync(),
+                DriverSetup.ConfigureChrome().WithMutex(true).SetUpAsync(),
+                DriverSetup.ConfigureChrome().WithMutex(true).SetUpAsync());
+
+            results.Should().HaveCount(3);
+            results[0].Should().BeEquivalentTo(results[1]).And.BeEquivalentTo(results[2]);
+            AssertDriverIsSetUp(results[0], BrowserNames.Chrome);
+            AssertVersionCache(BrowserNames.Chrome);
+        }
+
         private static void AssertAutoSetUpDriverResults(
             IEnumerable<DriverSetupResult> setupResults,
             IEnumerable<string> browserNames)